Fix issue with "N" confusing the parser
Problem: Commands like `M23` and `M33` must use lowercase in their
paths. Even the M117 command must avoid the letter `N`.

Solution: Instead of finding the “N” parameter anywhere in the command,
only look for it in the first parameter. This should be compliant with
all hosts and other software generating G-Code with line numbers at
this time.