Print Counter
Adds the function to store the number of prints in the EEPROM . Is based
in the code of @iashwinprasada with a few changes to make it work
independently if you have or not enable the EEPROM Settings.

Also I did a few changes in the way of implement this function with the
lcd, now is ready to add this feature in different languages currently
in (English and Spanish) need to translate the new menu
(MSG_PRINT_COUNTER, MSG_PRINTS).

The new Gcodes are

M505 G - Gets the current Print Count
M505 F - Increase the print counter by 1 (Indicates another Print has
Finished)
M505 C - Clears the Print Count

And you can check the number of prints with the lcd in:

Menu\Control\Print Counter

- Remove the automatic call for Increase the print count when prints
from SD
- Change EEPROM_VERSION to V22
- Adds the Option to Enable Print Counter in configuration file