Revising Configuration file structure
These file present the idea of how the configuration
variables can be reorganized to avoid the huge
amount of duplication and extraneous material found
in the present “flat” scheme that expects EVERY
configuration to provide a template that can be used
to build any possible configuration.

Instead, those template items can be moved into
the Documentation, hopefully with additional
explanation.

The resulting configuration files will be stripped
down to the items which are an exception to some
more general configuration.

Ultimately, all configurations will inherit from
a default configuration.